Fix to the content in Accessibility #53 and Help #54 Made the search results table responsive and it is now grouped by GP practice As per accessibility changes cited in #44, the following have been done:-
Sign in screen - move the Sign In button so that it is in or just below the yellow box containing the instructions. - DONE

If both Provider and Consumer ODS codes are not found, provide a separate message for each one and say which it is. - DONE

Yellow boxes are classed as 'Warning callouts' so not really suitable to display the provider and consumer practice names/addresses. Replace it with a Card - Feature Card. - DONE

The practice names and addresses (which are in the yellow box) need to be separated out, with headings of 'Searching at' (followed by the Provider name and address) and 'Searching on behalf of' (followed by the Consumer name and address) - DONE

Move the ODS input boxes and date drop-down so that they are directly below their labels. Move the Search for free slots and Clear buttons so that they are also directly below the labels/input boxes. Clear button to stay side-by-side with Search for free slots button - DONE

Move the details of Branch Surgeries above the column headings, so that the branch details are the heading for the table. Treat each branch surgery as a separate table. with its own column headings. - DONE